fix: modulaziração PF (individualPerson), PIS/RG/CPF privados, feat:serializer
