+++ registry/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 from .models import Municipality, Person, Address, IndividualPerson, BusinessPerson
+from .models.individualPerson import IndividualPerson
 
 
 class MunicipalitySerializer(serializers.ModelSerializer):
@@ -17,11 +18,28 @@ class AddressSerializer(serializers.ModelSerializer):
         model = Address
         fields = '__all__'
 
-class IndividualSerializer(PersonSerializer):  # Herda de PersonSerializer
+# class IndividualSerializer(PersonSerializer):  # Herda de PersonSerializer
+#     class Meta:
+#         model = IndividualPerson
+#         fields = '__all__'
+        
+class IndividualSerializer(PersonSerializer):  
     class Meta:
         model = IndividualPerson
         fields = '__all__'
 
+    def to_representation(self, instance):
+        data = super().to_representation(instance)
+        request = self.context.get('request')
+
+        # Exemplo: ocultar CPF e RG para usuários não autenticados
+        if request and not request.user.is_authenticated:
+            data.pop('cpf', None)
+            data.pop('rg', None)
+            data.pop('pis', None)
+
+        return data
+    
 
 class BusinessSerializer(PersonSerializer):  # Herda de PersonSerializer
     class Meta:
